// src/types/supabase.ts
export interface definitions {
  app_users: {
    /**
     * Note:
     * This is a Primary Key.<pk/>
     * This is a Foreign Key to `users.id`.<fk table='users' column='id'/>
     */
    id: string;
    full_name?: string;
    avatar_url?: string;
    billing_address?: string;
    payment_method?: string;
  };
  customers: {
    /**
     * Note:
     * This is a Primary Key.<pk/>
     * This is a Foreign Key to `users.id`.<fk table='users' column='id'/>
     */
    id: string;
    stripe_customer_id?: string;
  };
  jobs: {
    id: number;
    /**
     * Note:
     * This is a Foreign Key to `app_users.id`.<fk table='app_users' column='id'/>
     */
    user_id?: string;
    company_name: string;
    company_email?: string;
    company_number?: string;
    title: string;
    description: string;
    salary_type:
    | "annual_package"
    | "anual_commission"
    | "comission_only"
    | "hourly"
    | "project";
    work_type:
    | "casual"
    | "contract_temp"
    | "full_time"
    | "part_time"
    | "project";
    salary_min?: number;
    salary_max?: number;
    created_at: string;
    updated_at: string;
    expires_on: string;
  };
  stripe_prices: {
    /**
     * Note:
     * This is a Primary Key.<pk/>
     */
    id: string;
    /**
     * Note:
     * This is a Foreign Key to `stripe_products.id`.<fk table='stripe_products' column='id'/>
     */
    product_id?: string;
    active?: boolean;
    description?: string;
    unit_amount?: number;
    currency?: string;
    type?: "one_time" | "recurring";
    interval?: "day" | "week" | "month" | "year";
    interval_count?: number;
    trial_period_days?: number;
    metadata?: string;
  };
  stripe_products: {
    /**
     * Note:
     * This is a Primary Key.<pk/>
     */
    id: string;
    active?: boolean;
    name?: string;
    description?: string;
    image?: string;
    metadata?: string;
  };
  stripe_subscriptions: {
    id: string;
    /**
     * Note:
     * This is a Foreign Key to `users.id`.<fk table='users' column='id'/>
     */
    user_id: string;
    status?:
    | "trialing"
    | "active"
    | "canceled"
    | "incomplete"
    | "incomplete_expired"
    | "past_due"
    | "unpaid";
    metadata?: string;
    /**
     * Note:
     * This is a Foreign Key to `stripe_prices.id`.<fk table='stripe_prices' column='id'/>
     */
    price_id?: string;
    quantity?: number;
    cancel_at_period_end?: boolean;
    created: string;
    current_period_start: string;
    current_period_end: string;
    ended_at?: string;
    cancel_at?: string;
    canceled_at?: string;
    trial_start?: string;
    trial_end?: string;
  };
}